% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DE.R
\name{DE}
\alias{DE}
\alias{UniPro}
\title{Differential Evolution}
\usage{
DE(n, m, s = n, p = 15L, NP = 100L, itermax = 1000L, pMut = 0.2, pCR = 0.3,
          pGBest = 0.9, replicates = 1L, seed = sample(1e7,1), ncores = NULL,
          method = c("UniPro", "MaxPro"))

UniPro(n, m, s = n, NP = 100L, itermax = 1000L, pMut = 0.2, pCR = 0.3,
       pGBest = 0.9, replicates = 1L, seed = sample(1e7,1), ncores = NULL)
}
\arguments{
\item{n}{number of rows(observations)}

\item{m}{number of columns(variables/factors)}

\item{s}{number of levels. Default = n}

\item{NP}{initial population size. Default = 100}

\item{itermax}{maximum number of iterations. Default = 1000}

\item{pMut}{the probability of mutation}

\item{pCR}{Probability of CrossOver}

\item{pGBest}{Probability of using global best}

\item{replicates}{number of replications. Determines the output}

\item{seed}{seed to be set for reproducibility}

\item{ncores}{number f cores to be used for parallelization/multi-threading.}

\item{p}{the exponent for method maximinLHD. default is 15}

\item{method}{The criterion to be optimized. Only UniPro, MaxPro, maximinLHD are implemented}
}
\value{
A list with the phi value, the final design and total computational time taken:
\item{bestX}{The final UniPro Design. Only returned when replicates = 1}
\item{optValues}{The phi value of the final Design. A vector of length replicates}
\item{timeTaken}{Total computational time taken}
}
\description{
DE: A function to construct Designs using various criterions
algorithm

UniPro: A function to construct a Uniform Projection Design using Differential Evolution
algorithm
}
\examples{
DE(10, 3) # uses UniPro by default
DE(30, 3, replicates=10, method = 'MaxiPro')
UniPro(10, 3)
UniPro(30, 3, replicates=10)

}
