% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DE.R
\name{maximin}
\alias{maximin}
\title{Maximin-Distance Latin Hypercube Designs}
\usage{
maximinLHD(n, m, p = 15L, NP = 100L, itermax = 1000L, pMut = 0.2, pCR = 0.3,
       pGBest = 0.9, replicates = 1L, seed = sample(1e7,1), ncores = NULL)
}
\description{
maximinLHD: A function to construct  Maximin-Distance Latin Hypercube Designs using Differential Evolution
algorithm
}
\examples{
maximinLHD(10, 3)
maximinLHD(30, 3, replicates=10)

}
