% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DE.R
\name{maximinmeasure}
\alias{maximinmeasure}
\title{maxminmeasure}
\usage{
maximinmeasure(X)
}
\description{
A function that computes the maximin criterion value for a given design.
(Morris and Mitchell 1995)
}
\examples{
maximinmeasure(replicate(3, sample(10)))

}
