% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DE.R
\name{maxpromeasure}
\alias{maxpromeasure}
\title{maxpromeasure}
\usage{
maxpromeasure(X)
}
\description{
A function that computes the MaxPro criterion value for a given design.
Equation 5 of the  Maximum projection designs for computer experiments
paper BY V. ROSHAN JOSEPH,EVRENGUL
}
\examples{
maxpromeasure(replicate(3, sample(10)))

}
