% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UniPro.R
\name{phi}
\alias{phi}
\title{phi}
\usage{
phi_fortranR(X, s = nrow(X))
}
\arguments{
\item{X}{Design}

\item{s}{number of levels for the design X}
}
\value{
the phi value.
}
\description{
A function that computes the UniPro criterion value for a balanced design.
Equation 5.1 and 5.2 of the Uniform Projection Designs paper by Sun, Wang and Xu.
}
\examples{
phi(replicate(3, sample(10)))

}
