% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UniPro.R
\name{UniPro}
\alias{UniPro}
\title{Uniform Projection}
\usage{
UniPro(n, m, s = n, NP = 100, itermax = 1000,
              pMut =0.2, pCR = 0.3, pGBest = 0.9, replicates = 1)
}
\arguments{
\item{n}{number of rows(observations)}

\item{m}{number of columns(variables/factors)}

\item{s}{number of levels. Default = n}

\item{NP}{initial population size. Default = 100}

\item{itermax}{maximum number of iterations. Default = 1000}

\item{pMut}{the probability of mutation}

\item{pCR}{Probability of CrossOver}

\item{pGBest}{Probability of using global best}

\item{replicates}{number of replications. Determines the output}
}
\value{
A list with the phi value, the final design and total computational time taken:
\item{X}{The final UniPro Design. Only returned when replicates = 1}
\item{val}{The phi value of the final Design. A vector of length replicates}
\item{time_taken}{Total computational time taken}
}
\description{
A function to construct a Uniform Projection Design using Differential Evolution
algorithm
}
\examples{
UniPro(10, 3)
UniPro(30, 3, replicates=10)


}
