% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DE.R
\name{unipromeasure}
\alias{unipromeasure}
\title{unipromeasure}
\usage{
unipromeasure(X)
}
\arguments{
\item{s}{number of levels. Default = n}

\item{X}{Design}

\item{method}{The criterion to be optimized. Only UniPro, MaxPro, maximinLHD are implemented}
}
\value{
measure value.
}
\description{
A function that computes the UniPro criterion value for a balanced design.
Equation 5.1 and 5.2 of the Uniform Projection Designs paper by Sun, Wang and Xu.
}
\examples{
unipromeasure(replicate(3, sample(10)))

}
